bug in types

+++ bee_system_mast/photoresult.py
@@ -7,7 +7,7 @@ class PhotoResult():
         print(flash)
         print("Direction: ")
         print(direction)
-        self.img = img
+        self.img = img.astype(np.double)
         self.flash = flash
         self.direction = direction
         self.time = time.time()
